pyproject.toml
[tool.poetry]
name = "petshop-back"
version = "0.1.0"
description = ""
authors = ["Juan Ernesto Sierra <juersierra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
pymongo = {extras = ["srv"], version = "^4.8.0"}
pydantic-mongo = "^2.3.0"
fastapi-jwt = {extras = ["authlib"], version = "^0.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.2.0"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
